requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
crispy-forms-foundation==0.8.0
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-countries==7.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-qr-code==2.1.0
django-session-timeout==0.1.0
django-tables2==2.3.4
django-wkhtmltopdf==3.3.0
django-xhtml2pdf==0.0.4
future==0.18.2
html5lib==1.1
Pillow==8.1.0
PyMySQL==1.0.2
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
python-bidi==0.4.2
pytz==2021.1
qrcode==6.1
reportlab==3.5.59
segno==1.3.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.5
